PWM code working, canbus interference

+++ STM32_Firmware/Libraries/Drone_Can/can_node/Inc/can_battery.h
@@ -5,4 +5,4 @@
 #include <can_node.h>
 
 
-int16_t send_battery_info(uint16_t voltage, uint16_t current);
+int16_t send_battery_info(float voltage, float current);
